shift get_estimator_object to Causal Estimator
