+++ dowhy/causal_refuters/dummy_outcome_refuter.py
@@ -6,6 +6,7 @@ import logging
 
 from dowhy.causal_refuter import CausalRefutation
 from dowhy.causal_refuter import CausalRefuter
+from dowhy.causal_estimator import CausalEstimator
 
 class DummyOutcomeRefuter(CausalRefuter):
     """Refute an estimate by replacing the outcome with a randomly generated variable.
@@ -93,7 +94,7 @@ class DummyOutcomeRefuter(CausalRefuter):
         # Sanity check the data
         self.logger.debug(new_data[0:10])
 
-        new_estimator = self.get_estimator_object(new_data, identified_estimand, self._estimate)
+        new_estimator = CausalEstimator.get_estimator_object(new_data, identified_estimand, self._estimate)
         new_effect = new_estimator.estimate_effect()
         sample_estimates[index] = new_effect.value
